Change the first example to be complete X3D file. Not much more code, and is more impressive this way, I think.
Do not use the term "plugging point", just use "plug", consistently.

Mention "The node specification below is presented in the same style as in the X3D specification", to clearly indicate that X3D knowledge in probably needed at this point.

Mention in text "Inside the \texttt{Effect} node a number of uniform values may be defined...". This is cleaner than previous "fieldName xxx" shortcut, and makes another ref to X3D shaders component, which is important base.

Remove whole "We have also considered and rejected a different approach to use plugs" paragraph. This is just not really needed. We explain earlier that our approach is cool (separate compilation units), so no need to explain now that other approach is not cool because it doesn't give clean compilation units... Makes space.

git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/papers@9486 afe8dcc5-9094-451b-8773-5f495042a737